--- website-nextjs/src/app/assets/icons/index.ts
@@ -1,13 +1,0 @@
-export { default as ArrowUpIcon } from "./arrow-up.svg"
-export { default as CircleOutlineIcon } from "./circle-outline.svg"
-export { default as CloseIcon } from "./close.svg"
-export { default as ForkIcon } from "./fork-icon.svg"
-export { default as GithubIcon } from "./github.svg"
-export { default as HomeIcon } from "./home.svg"
-export { default as LampOnIcon } from "./lamp-on.svg"
-export { default as LayoutGroupIcon } from "./layout-group.svg"
-export { default as MainIcon } from "./main.svg"
-export { default as MenuIcon } from "./menu.svg"
-export { default as StarIcon } from "./star.svg"
-export { default as UserIcon } from "./user.svg"
-export { default as UsersAltIcon } from "./users-alt.svg"
