+++ website-nextjs/src/redux/store.ts
@@ -1,0 +1,22 @@
+import { createStore, applyMiddleware } from 'redux';
+import {thunk} from 'redux-thunk'
+import { composeWithDevTools } from '@redux-devtools/extension';
+import { createWrapper } from "next-redux-wrapper";
+import rootReducer from './rootReducers';
+
+// initial states here
+const initalState = {};
+
+// middleware
+const middleware = [thunk];
+
+export const store = createStore(
+    rootReducer,
+    initalState,
+    composeWithDevTools(applyMiddleware(...middleware))
+);
+
+// assigning store to next wrapper
+const makeStore = () => store;
+
+export const wrapper = createWrapper(makeStore);
